requirements.txt
python-dotenv
requests
pandas
numpy
plotly
pyarrow
scipy
dash